진행사항 업데이트 -> 서버에 맞게 변경 필요

+++ LOD_Server/LOD_Server/Code/00.Global/01.Utility/04.WayFinder/00.Node/Node.cpp
@@ -5,7 +5,7 @@
 /// 목적: 길찾기 알고리즘에서 사용할 노드 정의
 /// 최종 수정자:  김나단
 /// 수정자 목록:  김나단
-/// 최종 수정 날짜: 2018-04-20
+/// 최종 수정 날짜: 2018-07-03
 /// </summary>
 
 ////////////////////////////////////////////////////////////////////////
@@ -25,15 +25,15 @@ const POINT CNode::GetDividIndex() const
 {
 	POINT dividIndex;
 
-	dividIndex.x = m_position.x / m_size.x;
-	dividIndex.y = m_position.y / m_size.y;
+	dividIndex.x = static_cast<long>(m_position.x / m_size.x);
+	dividIndex.y = static_cast<long>(m_position.y / m_size.y);
 
 	return dividIndex;
 }
 
 int CNode::GetDistanceSquareWithPosition(const XMFLOAT2 &position)
 {
-	return  (position.x - m_position.x) * (position.x - m_position.x) + (position.y - m_position.y) * (position.y - m_position.y);
+	return  static_cast<int>((position.x - m_position.x) * (position.x - m_position.x) + (position.y - m_position.y) * (position.y - m_position.y));
 }
 
 
@@ -41,9 +41,9 @@ int CNode::GetDistanceSquareWithPosition(const XMFLOAT2 &position)
 // 내부 함수
 int CNode::PositionToIndex(XMFLOAT2 position)
 {
-	int x = position.x / m_size.x;
-	int y = position.y / m_size.y;
-	int xMax = TERRAIN_SIZE_WIDTH / m_size.x + 1;
+	int x = static_cast<int>(position.x / m_size.x);
+	int y = static_cast<int>(position.y / m_size.y);
+	int xMax = static_cast<int>(TERRAIN_SIZE_WIDTH / m_size.x + 1);
 
 	return  xMax * y + x;
 }
